add post folder and move all the post related component to it
