+++ src/app/dashboard/pages/post/details/post-details.component.scss
