+++ src/app/dashboard/pages/post/details/post-details.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
-import { Post } from '../models/post.model';
-import { PostService } from '../services/post.service';
+import { Post } from '../../models/post.model';
+import { PostService } from '../../services/post.service';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
