+++ src/app/dashboard/pages/post/manage/manage-create.component.scss
