+++ src/app/dashboard/pages/post/manage/manage-create.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { PostCreateComponent } from './post-create.component';
+import { PostCreateComponent } from './manage-create.component';
 
 describe('PostCreateComponent', () => {
   let component: PostCreateComponent;
